The/Missing/Page
This project was created under a series of several Computer and
Computation workshops. Using Processing I wanted to critique the
academic ways in which the history of Lebanon is taught to students.
The average Lebanese student learns in 4 academic years the history of
the world that includes  the prehistoric times, to the Lebanese
Independence in 1943. All the events that occurred after this date are
unclear and un-agreed on, regardless of the fact that during this time
a 30year civil war happened , followed by years of unstable situations
till this day. My approach to this topic is to urge the Lebanese
population to demand the right to have an understanding of their full
history in order for them to start building nation based on stable
grounds. This project Is called the missing page it is the random
history we could have since nothing is written down.